Fixed the physics terrain shape

+++ GoblinBrawl/GoblinBrawl/Floor.cpp
@@ -157,8 +157,8 @@ BYTE* Floor::GetRawHeightData( int gridSize, float heightScale, btScalar gridSpa
 	Pixel test0 = pixelData[0];
 	Pixel test1 = pixelData[1];
 	Pixel test2 = pixelData[2];
-	int texWidth = heightfieldMap.RowPitch/8; // 8 bytes per pixel
-	int gridToTexScale = texWidth/gridSize;
+	float texWidth = heightfieldMap.RowPitch/8; // 8 bytes per pixel
+	float gridToTexScale = texWidth/gridSize;
 	long nBytes = nElements * bytesPerElement;
 	BYTE * raw = new BYTE[nBytes];
 	btAssert( raw && "out of memory" );
@@ -166,9 +166,9 @@ BYTE* Floor::GetRawHeightData( int gridSize, float heightScale, btScalar gridSpa
 	for( int i = 0; i<gridSize; ++i ) {
 		for( int j = 0; j<gridSize; ++j ) {
 			//float value = 4.75;
-			int xTexCoord = (int)(j * gridToTexScale);
-			int yTexCoord = (int)(i * gridToTexScale);
-			int pixel = (int)(yTexCoord*texWidth+xTexCoord);
+			int xTexCoord = j * gridToTexScale;
+			int yTexCoord = i * gridToTexScale;
+			int pixel = yTexCoord*texWidth+xTexCoord;
 			float value = (float)(pixelData[pixel].r/65536.f)*heightScale;
 			switch( type ) {
 			case PHY_FLOAT:
